Moving client/server into similar docker builds - Switching between laptops with different distros was driving me nuts so this is a workaround - Tests are separated, AlsaTest builds now - Client does build, unsure how to deal with protoc warnings
