+++ cmake/CompilerWarnings.cmake
@@ -72,9 +72,7 @@ function(set_project_warnings project_name)
       -Wduplicated-branches # warn if if / else branches have duplicated code
       -Wlogical-op # warn about logical operations being used where bitwise were
                    # probably wanted
-      # TODO protoc-generated files trigger this warning, figure out a way to disable
-      #      it just for those files.
-      #-Wuseless-cast # warn if you perform a cast to the same type
+      -Wuseless-cast # warn if you perform a cast to the same type
   )
 
   if(MSVC)
